Implement autologging and data splitting function
Enabled autologging using mlflow and added a data split function in the training module to divide the dataframe into training and testing datasets. In addition, updated the `job.yml` to run `train.py` with specific parameters and added an explicit type hint for the `parse_args` function in `train.py`.