Update job trigger and install Azure CLI in workflow
Switched the manual trigger of an Azure Machine Learning job to trigger on push in the GitHub Actions workflow. In addition, the Azure CLI was added to enhance automation and allow for automated triggering of machine learning jobs on Azure. The 'ml' az extension is now installed as part of the updated workflow.